+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # cuML 0.8.0 (Date TBD)
 
 ## New Features
+- PR #515: Added Random Projection feature
 - PR #504: Contingency matrix ml-prim
 
 ## Improvements